putting the test into the appropriate file
